==> 03_State_popup.Rmd <==
---
title: "Popups"
author: "Diana Villasana Ocampo"
output:
   html_document:
      highlight: tango
      theme: flatly
      toc: yes
      toc_depth: 2
      toc_float:
        collapsed: yes
---
\usepackage{color}

```{=html}
<style>
code.r{
  font-size: 10px;
}
pre {
  font-size: 12px
}
</style>

<style>
body {
text-align: justify;
font-style: normal;
font-family: "Montserrat";
font-size: 12px
}
h1.title {
  font-size: 40px;
  color: #000D3B;
}
h1 {
  color: #B6854D;
}
h2 {
  color: #172984;
}
h3 {
  color: #172984;
}
</style>
```

```{=html}
<style>
.nav>li>a {
    position: relative;
    display: block;
    padding: 10px 15px;
    color: #0A2687;
}
.nav-pills>li.active>a, .nav-pills>li.active>a:hover, .nav-pills>li.active>a:focus {
    color: #ffffff;
    background-color: #09C2BC;
}
</style>
```

```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, cache = FALSE, cache.lazy = FALSE, 
                         eval = FALSE, class.source = "fold-show")
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
options(digits = 2, encoding = "UTF8")
```   
 

```{r, echo=FALSE}
rm(list = ls())
```

```{r, echo=FALSE}
setwd(here::here())
```


```{r}
require(dplyr)
require(forcats)
require(tibble)
require(sf)
require(sp)
require(stringr)
require(dplyr)
require(leaflet)
require(leafgl)
require(RColorBrewer)
```
# Base de datos   

```{r}
load("Output/IMC_2020.RData")
```


# Shapefiles 

```{r}
## Shapefiles
shape_colonias_1 <- geojsonio::geojson_read("Output/Shapes/colonias_1_json.geojson", what = "sp") 
shape_colonias_2 <- geojsonio::geojson_read("Output/Shapes/colonias_2_json.geojson", what = "sp") 
shape_colonias_3 <- geojsonio::geojson_read("Output/Shapes/colonias_3_json.geojson", what = "sp") 
shape_colonias_4 <- geojsonio::geojson_read("Output/Shapes/colonias_4_json.geojson", what = "sp") 
shape_colonias_5 <- geojsonio::geojson_read("Output/Shapes/colonias_5_json.geojson", what = "sp")
```


# Paleta de colores 

```{r}
paleta <- c("#13322B", #Muy alto
            "#086953", #Alto 
            "#4C9562", #Medio
            "#D1BB9E", #Bajo
            "#C3BA6D") # Muy bajo

mypalette <- leaflet::colorFactor(palette = paleta, 
                                   domain = forcats::fct_relevel(IMC_2020$GM_2020, c("Muy alto", "Alto", "Medio", "Bajo", "Muy bajo")), 
                                   na.color = "transparent")
```


## Popup `Muy alto`

```{r}
## Paleta de colores                         
p <- shape_colonias_1 %>% 
      leaflet() %>% 
       addTiles() %>% 
        setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
         addProviderTiles(providers$Esri.WorldImagery) %>%
          addPolygons(fillColor = ~mypalette(GM_2020))

fillColor <- p[["x"]][["calls"]][[3]][["args"]][[4]][["fillColor"]]

df_colonias <- shape_colonias_1@data %>%
                select(CVE_COL) %>%
                 inner_join(., IMC_2020 %>% 
                                 mutate(across(.cols = c(5:17,19), as.numeric)) %>%
                                  filter(GM_2020 %in% "Muy alto"),
                            by = c("CVE_COL"))

## Etiquetas del mapa a nivel municipal 
state_popup_colonia <- paste('<font size="3"; font-family: "Montserrat"><strong>',"Estado:",'</strong>','<strong>', df_colonias$NOM_ENT,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Municipio:",'</strong>','<strong>', df_colonias$NOM_MUN,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Colonia:",'</strong>','<strong>', df_colonias$NOM_COLONIA,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', formatC(df_colonias$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', df_colonias$GM_2020,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IM_2020, 2),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IMN_2020, 2),'</strong></font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 6 a 14 años que no asiste a la escuela:",'</strong>', round(df_colonias$P6A14NAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_colonias$SBASC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de sin derechohabiencia a los servicios de salud:",'</strong>', round(df_colonias$PSDSS, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_colonias$OVSDE, 2),'<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_colonias$OVSEE, 2),'</font>','<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_colonias$OVSAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_colonias$OVPT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas con algún nivel de hacinamiento:",'</strong>', round(df_colonias$OVHAC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin refrigerador:",'</strong>', round(df_colonias$OVSREF, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin internet:",'</strong>', round(df_colonias$OVSINT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin celular:",'</strong>', round(df_colonias$OVSCEL, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_colonia, "Output/PopUp/state_popup_colonia1.RDS")
```

## Popup `Alto`

```{r}
## Paleta de colores                         
p <- shape_colonias_2 %>% 
      leaflet() %>% 
       addTiles() %>% 
        setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
         addProviderTiles(providers$Esri.WorldImagery) %>%
          addPolygons(fillColor = ~mypalette(GM_2020))

fillColor <- p[["x"]][["calls"]][[3]][["args"]][[4]][["fillColor"]]

df_colonias <- shape_colonias_2@data %>%
                select(CVE_COL) %>%
                 inner_join(., IMC_2020 %>% 
                                mutate(across(.cols = c(5:17,19), as.numeric)) %>%
                                 filter(GM_2020 %in% "Alto"),
                    by = c("CVE_COL"))

## Etiquetas del mapa a nivel municipal 
state_popup_colonia <- paste('<font size="3"; font-family: "Montserrat"><strong>',"Estado:",'</strong>','<strong>', df_colonias$NOM_ENT,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Municipio:",'</strong>','<strong>', df_colonias$NOM_MUN,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Colonia:",'</strong>','<strong>', df_colonias$NOM_COLONIA,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', formatC(df_colonias$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', df_colonias$GM_2020,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IM_2020, 2),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IMN_2020, 2),'</strong></font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 6 a 14 años que no asiste a la escuela:",'</strong>', round(df_colonias$P6A14NAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_colonias$SBASC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de sin derechohabiencia a los servicios de salud:",'</strong>', round(df_colonias$PSDSS, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_colonias$OVSDE, 2),'<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_colonias$OVSEE, 2),'</font>','<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_colonias$OVSAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_colonias$OVPT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas con algún nivel de hacinamiento:",'</strong>', round(df_colonias$OVHAC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin refrigerador:",'</strong>', round(df_colonias$OVSREF, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin internet:",'</strong>', round(df_colonias$OVSINT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin celular:",'</strong>', round(df_colonias$OVSCEL, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_colonia, "Output/PopUp/state_popup_colonia2.RDS")
```


## Popup `Medio`

```{r}
## Paleta de colores                         
p <- shape_colonias_3 %>% 
      leaflet() %>% 
       addTiles() %>% 
        setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
         addProviderTiles(providers$Esri.WorldImagery) %>%
          addPolygons(fillColor = ~mypalette(GM_2020))

fillColor <- p[["x"]][["calls"]][[3]][["args"]][[4]][["fillColor"]]

df_colonias <- shape_colonias_3@data %>%
                select(CVE_COL) %>%
                 inner_join(., IMC_2020 %>% 
                                 mutate(across(.cols = c(5:17,19), as.numeric)) %>%
                                 filter(GM_2020 %in% "Medio"),
                               by = c("CVE_COL"))

## Etiquetas del mapa a nivel municipal 
state_popup_colonia <- paste('<font size="3"; font-family: "Montserrat"><strong>',"Estado:",'</strong>','<strong>', df_colonias$NOM_ENT,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Municipio:",'</strong>','<strong>', df_colonias$NOM_MUN,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Colonia:",'</strong>','<strong>', df_colonias$NOM_COLONIA,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', formatC(df_colonias$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', df_colonias$GM_2020,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IM_2020, 2),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IMN_2020, 2),'</strong></font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 6 a 14 años que no asiste a la escuela:",'</strong>', round(df_colonias$P6A14NAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_colonias$SBASC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de sin derechohabiencia a los servicios de salud:",'</strong>', round(df_colonias$PSDSS, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_colonias$OVSDE, 2),'<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_colonias$OVSEE, 2),'</font>','<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_colonias$OVSAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_colonias$OVPT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas con algún nivel de hacinamiento:",'</strong>', round(df_colonias$OVHAC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin refrigerador:",'</strong>', round(df_colonias$OVSREF, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin internet:",'</strong>', round(df_colonias$OVSINT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin celular:",'</strong>', round(df_colonias$OVSCEL, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_colonia, "Output/PopUp/state_popup_colonia3.RDS")
```

## Popup `Bajo`

```{r}
## Paleta de colores                         
p <- shape_colonias_4 %>% 
      leaflet() %>% 
       addTiles() %>% 
        setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
         addProviderTiles(providers$Esri.WorldImagery) %>%
          addPolygons(fillColor = ~mypalette(GM_2020))

fillColor <- p[["x"]][["calls"]][[3]][["args"]][[4]][["fillColor"]]

df_colonias <- shape_colonias_4@data %>%
                select(CVE_COL) %>%
                 inner_join(., IMC_2020 %>% 
                                mutate(across(.cols = c(5:17,19), as.numeric)) %>%
                                 filter(GM_2020 %in% "Bajo"),
                               by = c("CVE_COL"))

## Etiquetas del mapa a nivel municipal 
state_popup_colonia <- paste('<font size="3"; font-family: "Montserrat"><strong>',"Estado:",'</strong>','<strong>', df_colonias$NOM_ENT,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Municipio:",'</strong>','<strong>', df_colonias$NOM_MUN,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Colonia:",'</strong>','<strong>', df_colonias$NOM_COLONIA,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', formatC(df_colonias$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', df_colonias$GM_2020,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IM_2020, 2),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IMN_2020, 2),'</strong></font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 6 a 14 años que no asiste a la escuela:",'</strong>', round(df_colonias$P6A14NAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_colonias$SBASC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de sin derechohabiencia a los servicios de salud:",'</strong>', round(df_colonias$PSDSS, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_colonias$OVSDE, 2),'<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_colonias$OVSEE, 2),'</font>','<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_colonias$OVSAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_colonias$OVPT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas con algún nivel de hacinamiento:",'</strong>', round(df_colonias$OVHAC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin refrigerador:",'</strong>', round(df_colonias$OVSREF, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin internet:",'</strong>', round(df_colonias$OVSINT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin celular:",'</strong>', round(df_colonias$OVSCEL, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_colonia, "Output/PopUp/state_popup_colonia4.RDS")
```

## Popup `Muy Bajo`

```{r}
## Paleta de colores                         
p <- shape_colonias_5 %>% 
      leaflet() %>% 
       addTiles() %>% 
        setView(lat = 23.634501, lng = -102.552784, zoom = 5) %>%
         addProviderTiles(providers$Esri.WorldImagery) %>%
          addPolygons(fillColor = ~mypalette(GM_2020))

fillColor <- p[["x"]][["calls"]][[3]][["args"]][[4]][["fillColor"]]

df_colonias <- shape_colonias_5@data %>%
                select(CVE_COL) %>%
                 inner_join(., IMC_2020 %>% 
                                mutate(across(.cols = c(5:17,19), as.numeric)) %>%
                                 filter(GM_2020 %in% "Muy bajo"),
                               by = c("CVE_COL"))

## Etiquetas del mapa a nivel municipal 
state_popup_colonia <- paste('<font size="3"; font-family: "Montserrat"><strong>',"Estado:",'</strong>','<strong>', df_colonias$NOM_ENT,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Municipio:",'</strong>','<strong>', df_colonias$NOM_MUN,'</strong></font>','<br/>',
                          '<font size="2.5"; font-family: "Montserrat"><strong>',"Colonia:",'</strong>','<strong>', df_colonias$NOM_COLONIA,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Población Total:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', formatC(df_colonias$POB_TOT, format = "f", big.mark = " ", digits = 0),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Grado de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', df_colonias$GM_2020,'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IM_2020, 2),'</strong></font>','<br/>',
                          '<font size="2"; font-family: "Montserrat"><strong>',"Índice de marginación normalizado:",'</strong></font>','<font size="2"; font-family: "Montserrat"; color=', fillColor,'><strong>', round(df_colonias$IMN_2020, 2),'</strong></font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 6 a 14 años que no asiste a la escuela:",'</strong>', round(df_colonias$P6A14NAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de 15 años o más sin educación básica:",'</strong>', round(df_colonias$SBASC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Población de sin derechohabiencia a los servicios de salud:",'</strong>', round(df_colonias$PSDSS, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin drenaje ni excusado:",'</strong>', round(df_colonias$OVSDE, 2),'<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin energía eléctrica:",'</strong>', round(df_colonias$OVSEE, 2),'</font>','<br/>', 
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin agua entubada:",'</strong>', round(df_colonias$OVSAE, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares con piso de tierra:",'</strong>', round(df_colonias$OVPT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas con algún nivel de hacinamiento:",'</strong>', round(df_colonias$OVHAC, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin refrigerador:",'</strong>', round(df_colonias$OVSREF, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin internet:",'</strong>', round(df_colonias$OVSINT, 2),'</font>','<br/>',
                          '<font size="1.5"; font-family: "Montserrat"><strong>',"% Ocupantes en viviendas paticulares sin celular:",'</strong>', round(df_colonias$OVSCEL, 2),'</font>'
) %>% 
  lapply(htmltools::HTML)

saveRDS(state_popup_colonia, "Output/PopUp/state_popup_colonia5.RDS")
```
